Project wrap-up: All functionalities tested and working

+++ app/src/main/java/com/example/myweather/weather_data/Day.kt
@@ -21,5 +21,5 @@ data class Day(
     val totalprecip_in: Double,
     val totalprecip_mm: Double,
     val totalsnow_cm: Int,
-    val uv: Int
+    val uv: Double
 )
